+++ trunk/OgreMain/src/OgreSubMesh.cpp
@@ -180,7 +180,7 @@ namespace Ogre {
 		ushort currNumFaces;
 		ushort *currFaces;
 
-		if (lodIndex > 0 && lodIndex-1 < mLodFaceList.size())
+		if (lodIndex > 0 && static_cast< size_t >( lodIndex - 1 ) < mLodFaceList.size())
 		{
 			// lodIndex - 1 because we don't store full detail version in mLodFaceList
 			currNumFaces = mLodFaceList[lodIndex-1].numIndexes / 3;
